pyproject.toml
[tool.poetry]
name = "campfirechat"
version = "0.1.0"
description = "chatGPT RPG prompting"
authors = ["hortenbach <gina.hortenbach@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.3.1"

[tool.poetry.group.dev.dependencies]
openai = "^0.27.7"
python-decouple = "^3.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
